Update showing textbox functions
Since I've switched to using Visual Studio for this project, the file is going to be compliant with the requirements of the MSVC++ compiler.
This is the reason the former int main() function has been changed to int CALLBACK WinMain(...) since the new linker seems to require a WinMain() function in a windows subsystem-based linking. We could continue using int main() by specifying a console subsystem but that would necessarily spawn a console window which we could hide via a ShowWindow but it's not an elegant solution - and the console blink would be clearly visible.
And the header file is for windows anyway. So, this is how it stands.